[FEAT/CH-13] 카카오 로그인 연결 및 스탬프 api연결

+++ app/src/main/java/com/example/check/model/bbti/BBTIResultRequest.java
@@ -1,20 +1,22 @@
 package com.example.check.model.bbti;
 
+import android.util.Log;
+
 public class BBTIResultRequest {
-    private int userid;
+    private Long userid;
     private int bbti;
 
-    public BBTIResultRequest(int userid, int bbti) {
+    public BBTIResultRequest(Long userid, int bbti) {
         this.userid = userid;
         this.bbti = bbti;
     }
 
     // Getters and setters
-    public int getUserid() {
+    public Long getUserid() {
         return userid;
     }
 
-    public void setUserid(int userid) {
+    public void setUserid(Long userid) {
         this.userid = userid;
     }
