+++ app/src/main/java/com/example/check/model/home/RecentLibrariesWrapper.java
@@ -5,16 +5,16 @@ import com.google.gson.annotations.SerializedName;
 import java.util.List;
 
 public class RecentLibrariesWrapper {
-    private int userid;
+    private Long userid;
 
     @SerializedName("recent_libraries")
     private List<RecentLibrary> recentLibraries;
 
-    public int getUserid() {
+    public Long getUserid() {
         return userid;
     }
 
-    public void setUserid(int userid) {
+    public void setUserid(Long userid) {
         this.userid = userid;
     }
